+++ src/gevent/_abstract_linkable.py
@@ -26,40 +26,44 @@ class AbstractLinkable(object):
     # Encapsulates the standard parts of the linking and notifying
     # protocol common to both repeatable events (Event, Semaphore) and
     # one-time events (AsyncResult).
+    #
+    # TODO: As of gevent 1.5, we use the same datastructures and almost
+    # the same algorithm as Greenlet. See about unifying them more.
 
-    __slots__ = ('hub', '_links', '_notifier', '_notify_all', '__weakref__')
+    __slots__ = (
+        'hub',
+        '_links',
+        '_notifier',
+        '_notify_all',
+        '__weakref__'
+    )
 
     def __init__(self):
         # Before this implementation, AsyncResult and Semaphore
         # maintained the order of notifications, but Event did not.
 
-        # In gevent 1.3, before Semaphore extended this class,
-        # that was changed to not maintain the order. It was done because
+        # In gevent 1.3, before Semaphore extended this class, that
+        # was changed to not maintain the order. It was done because
         # Event guaranteed to only call callbacks once (a set) but
-        # AsyncResult had no such guarantees.
+        # AsyncResult had no such guarantees. When Semaphore was
+        # changed to extend this class, it lost its ordering
+        # guarantees. Unfortunately, that made it unfair. There are
+        # rare cases that this can starve a greenlet
+        # (https://github.com/gevent/gevent/issues/1487) and maybe
+        # even lead to deadlock (not tested).
 
-        # Semaphore likes to maintain order of callbacks, though,
-        # so when it was added we went back to a list implementation
-        # for storing callbacks. But we want to preserve the unique callback
-        # property, so we manually check.
-
-        # We generally don't expect to have so many waiters (for any of those
-        # objects) that testing membership and removing is a bottleneck.
-
-        # In PyPy 2.6.1 with Cython 0.23, `cdef public` or `cdef
-        # readonly` or simply `cdef` attributes of type `object` can appear to leak if
-        # a Python subclass is used (this is visible simply
-        # instantiating this subclass if _links=[]). Our _links and
-        # _notifier are such attributes, and gevent.thread subclasses
-        # this class. Thus, we carefully manage the lifetime of the
-        # objects we put in these attributes so that, in the normal
-        # case of a semaphore used correctly (deallocated when it's not
-        # locked and no one is waiting), the leak goes away (because
-        # these objects are back to None). This can also be solved on PyPy
-        # by simply not declaring these objects in the pxd file, but that doesn't work for
-        # CPython ("No attribute...")
-        # See https://github.com/gevent/gevent/issues/660
-        self._links = set()
+        # So in gevent 1.5 we go back to maintaining order. But it's
+        # still important not to make duplicate calls, and it's also
+        # important to avoid O(n^2) behaviour that can result from
+        # naive use of a simple list due to the need to handle removed
+        # links in the _notify_links loop. Cython has special support for
+        # built-in sets, lists, and dicts, but not ordereddict. Rather than
+        # use two data structures, or a dict({link: order}), we simply use a
+        # list and remove objects as we go, keeping track of them so as not to
+        # have duplicates called. This makes `unlink` O(n), but we can avoid
+        # calling it in the common case in _wait_core (even so, the number of
+        # waiters should usually be pretty small)
+        self._links = []
         self._notifier = None
         # This is conceptually a class attribute, defined here for ease of access in
         # cython. If it's true, when notifiers fire, all existing callbacks are called.
@@ -95,13 +99,15 @@ class AbstractLinkable(object):
         """
         if not callable(callback):
             raise TypeError('Expected callable: %r' % (callback, ))
-
-        self._links.add(callback)
+        self._links.append(callback)
         self._check_and_notify()
 
     def unlink(self, callback):
         """Remove the callback set by :meth:`rawlink`"""
-        self._links.discard(callback)
+        try:
+            self._links.remove(callback)
+        except ValueError:
+            pass
 
         if not self._links and self._notifier is not None:
             # If we currently have one queued, de-queue it.
@@ -110,37 +116,42 @@ class AbstractLinkable(object):
             # But we can't set it to None in case it was actually running.
             self._notifier.stop()
 
-
     def _notify_links(self):
         # We release self._notifier here. We are called by it
         # at the end of the loop, and it is now false in a boolean way (as soon
         # as this method returns).
         notifier = self._notifier
-        # We were ready() at the time this callback was scheduled;
-        # we may not be anymore, and that status may change during
-        # callback processing. Some of our subclasses will want to
-        # notify everyone that the status was once true, even though not it
-        # may not be anymore.
-        todo = set(self._links)
+        # Early links are allowed to remove later links, and links
+        # are allowed to add more links.
+        #
+        # We were ready() at the time this callback was scheduled; we
+        # may not be anymore, and that status may change during
+        # callback processing. Some of our subclasses (Event) will
+        # want to notify everyone who was registered when the status
+        # became true that it was once true, even though it may not be
+        # anymore. In that case, we must not keep notifying anyone that's
+        # newly added after that, even if we go ready again.
+        final_link = self._links[-1]
+        only_while_ready = not self._notify_all
+        done = set() # of ids
         try:
-            for link in todo:
-                if not self._notify_all and not self.ready():
+            while self._links: # remember this can be mutated
+                if only_while_ready and not self.ready():
                     break
 
-                if link not in self._links:
-                    # Been removed already by some previous link. OK, fine.
+                link = self._links.pop(0) # Cython optimizes using list internals
+                id_link = id(link)
+                if id_link in done:
                     continue
+                done.add(id_link)
                 try:
                     link(self)
                 except: # pylint:disable=bare-except
-                    # We're running in the hub, so getcurrent() returns
-                    # a hub.
-                    self.hub.handle_error((link, self), *sys.exc_info()) # pylint:disable=undefined-variable
-                finally:
-                    if getattr(link, 'auto_unlink', None):
-                        # This attribute can avoid having to keep a reference to the function
-                        # *in* the function, which is a cycle
-                        self.unlink(link)
+                    # We're running in the hub, errors must not escape.
+                    self.hub.handle_error((link, self), *sys.exc_info())
+
+                if link is final_link:
+                    break
         finally:
             # We should not have created a new notifier even if callbacks
             # released us because we loop through *all* of our links on the
@@ -148,11 +159,11 @@ class AbstractLinkable(object):
             assert self._notifier is notifier
             self._notifier = None
 
-        # Our set of active links changed, and we were told to stop on the first
-        # time we went unready. See if we're ready, and if so, go around
-        # again.
-        if not self._notify_all and todo != self._links:
-            self._check_and_notify()
+        # Now we may be ready or not ready. If we're ready, which
+        # could have happened during the last link we called, then we
+        # must have more links than we started with. We need to schedule the
+        # wakeup.
+        self._check_and_notify()
 
     def _wait_core(self, timeout, catch=Timeout):
         # The core of the wait implementation, handling
@@ -161,23 +172,25 @@ class AbstractLinkable(object):
         # Returns a true value if the wait succeeded without timing out.
         switch = getcurrent().switch # pylint:disable=undefined-variable
         self.rawlink(switch)
-        try:
-            with Timeout._start_new_or_dummy(timeout) as timer:
-                try:
-                    if self.hub is None:
-                        self.hub = get_hub()
-                    result = self.hub.switch()
-                    if result is not self: # pragma: no cover
-                        raise InvalidSwitchError('Invalid switch into Event.wait(): %r' % (result, ))
-                    return True
-                except catch as ex:
-                    if ex is not timer:
-                        raise
-                    # test_set_and_clear and test_timeout in test_threading
-                    # rely on the exact return values, not just truthish-ness
-                    return False
-        finally:
-            self.unlink(switch)
+        with Timeout._start_new_or_dummy(timeout) as timer:
+            try:
+                if self.hub is None:
+                    self.hub = get_hub()
+                result = self.hub.switch()
+                if result is not self: # pragma: no cover
+                    raise InvalidSwitchError('Invalid switch into Event.wait(): %r' % (result, ))
+                # If we got here, we were automatically unlinked already.
+                return True
+            except catch as ex:
+                self.unlink(switch)
+                if ex is not timer:
+                    raise
+                # test_set_and_clear and test_timeout in test_threading
+                # rely on the exact return values, not just truthish-ness
+                return False
+            except:
+                self.unlink(switch)
+                raise
 
     def _wait_return_value(self, waited, wait_success):
         # pylint:disable=unused-argument